Remove forward declaration of TwitchAccount in TwitchAccountManager
