+++ src/controllers/commands/builtin/twitch/ChatSettings.cpp
@@ -4,6 +4,7 @@
 #include "controllers/accounts/AccountController.hpp"
 #include "messages/MessageBuilder.hpp"
 #include "providers/twitch/api/Helix.hpp"
+#include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
 #include "util/FormatTime.hpp"
 #include "util/Helpers.hpp"
