+++ src/controllers/ignores/IgnoreController.cpp
@@ -2,6 +2,7 @@
 
 #include "common/QLogging.hpp"
 #include "controllers/ignores/IgnorePhrase.hpp"
+#include "providers/twitch/TwitchAccount.hpp"
 #include "singletons/Settings.hpp"
 
 namespace chatterino {
