+++ src/widgets/dialogs/ReplyThreadPopup.cpp
@@ -5,6 +5,7 @@
 #include "common/QLogging.hpp"
 #include "controllers/hotkeys/HotkeyController.hpp"
 #include "messages/MessageThread.hpp"
+#include "providers/twitch/TwitchAccount.hpp"
 #include "util/LayoutCreator.hpp"
 #include "widgets/helper/ChannelView.hpp"
 #include "widgets/helper/ResizingTextEdit.hpp"
